add some services and make some mock data , link frontend with backend , make a login and sinup auth, link pages with each other

+++ backend/src/main/java/com/example/backend/repository/DepartmentRepo.java
@@ -3,7 +3,7 @@ package com.example.backend.repository;
 import com.example.backend.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepo extends JpaRepository<Department,Integer> {
+public interface DepartmentRepo extends JpaRepository<Department, Long> {
 
     // TODO: Implement repository logic (JPA/Hibernate/etc.)
